refactor endpoint handlers
