+++ packages/mds-agency/request-handlers.ts
@@ -37,7 +37,8 @@ import {
   BoundingBox,
   VEHICLE_REASON
 } from '@mds-core/mds-types'
-import { badDevice } from './utils'
+import { badDevice, getVehicles, lower, writeTelemetry, badEvent, badTelemetry, getServiceArea } from './utils'
+import urls from 'url'
 
 export const getAllServiceAreas = async (req: AgencyApiRequest, res: AgencyApiResponse) => {
   try {
@@ -161,5 +162,372 @@ export const registerVehicle = async (req: AgencyApiRequest, res: AgencyApiRespo
       await log.error(providerName(res.locals.provider_id), 'register vehicle failed:', err)
       res.status(500).send(new ServerError())
     }
+  }
+}
+
+export const getVehicleById = async (req: AgencyApiRequest, res: AgencyApiResponse) => {
+  const { device_id } = req.params
+
+  const { cached } = req.query
+
+  const { provider_id } = res.locals
+
+  function finish(device: Device, event?: VehicleEvent, telemetry?: Recorded<Telemetry> | Telemetry): void {
+    if (device.provider_id !== provider_id) {
+      res.status(404).send({
+        error: 'not_found'
+      })
+      return
+    }
+    const composite: Partial<
+      Device & { prev_event?: string; updated?: Timestamp; gps?: Recorded<Telemetry>['gps'] }
+    > = {
+      ...device
+    }
+
+    if (event) {
+      composite.prev_event = event.event_type
+      composite.updated = event.timestamp
+      composite.status = (EVENT_STATUS_MAP[event.event_type as VEHICLE_EVENT] || 'unknown') as VEHICLE_STATUS
+    } else {
+      composite.status = VEHICLE_STATUSES.removed
+    }
+    if (telemetry) {
+      if (telemetry.gps) {
+        composite.gps = telemetry.gps
+      }
+    }
+    res.send(composite)
+  }
+
+  log.info(`/vehicles/${device_id}`, cached)
+  if (cached) {
+    try {
+      const device = await cache.readDevice(device_id)
+      const event = await cache.readEvent(device_id).catch(async err => {
+        await log.warn(err)
+        return undefined
+      })
+      const telemetry = await cache.readTelemetry(device_id).catch(async err => {
+        await log.warn(err)
+        return undefined
+      })
+      if (device) return finish(device, event, telemetry)
+    } catch (err) {
+      await log.warn(providerName(res.locals.provider_id), `fail GET /vehicles/${device_id}`)
+      await log.error(err)
+      res.status(404).send({
+        error: 'not_found'
+      })
+    }
+  } else {
+    try {
+      const device = await db.readDevice(device_id).catch(async err => {
+        await log.error(err)
+        res.status(404).send({
+          error: 'not_found'
+        })
+      })
+      const event = await db.readEvent(device_id).catch(async err => {
+        await log.warn(err)
+        return undefined
+      })
+      const telemetry = await db.readTelemetry(device_id)
+      if (device) return finish(device, event, telemetry[0])
+    } catch (err) {
+      await log.error(err)
+      res.status(500).send(new ServerError())
+    }
+  }
+}
+
+export const getVehiclesByProvider = async (req: AgencyApiRequest, res: AgencyApiResponse) => {
+  let { skip, take } = req.query
+  const PAGE_SIZE = 1000
+
+  skip = parseInt(skip) || 0
+  take = parseInt(take) || PAGE_SIZE
+
+  const url = urls.format({
+    protocol: req.get('x-forwarded-proto') || req.protocol,
+    host: req.get('host'),
+    pathname: req.path
+  })
+
+  const { provider_id } = res.locals
+
+  try {
+    const response = await getVehicles(skip, take, url, provider_id, req.query)
+    return res.status(200).send(response)
+  } catch (err) {
+    await log.error('getVehicles fail', err)
+    res.status(500).send(new ServerError())
+  }
+}
+
+export const updateVehicle = async (req: AgencyApiRequest, res: AgencyApiResponse) => {
+  const { device_id } = req.params
+
+  const { vehicle_id } = req.body
+
+  const update = {
+    vehicle_id
+  }
+
+  const { provider_id } = res.locals
+
+  async function fail(err: Error | string) {
+    /* istanbul ignore else cannot easily test server failure */
+    if (String(err).includes('not found')) {
+      res.status(404).send({
+        error: 'not_found'
+      })
+    } else if (String(err).includes('invalid')) {
+      res.status(400).send({
+        error: 'invalid_data'
+      })
+    } else if (!provider_id) {
+      res.status(404).send({
+        error: 'not_found'
+      })
+    } else {
+      await log.error(providerName(provider_id), `fail PUT /vehicles/${device_id}`, req.body, err)
+      res.status(500).send(new ServerError())
+    }
+  }
+
+  try {
+    const tempDevice = await db.readDevice(device_id, provider_id)
+    if (tempDevice.provider_id !== provider_id) {
+      await fail('not found')
+    } else {
+      const device = await db.updateDevice(device_id, provider_id, update)
+      await Promise.all([cache.writeDevice(device), stream.writeDevice(device)])
+      return res.status(201).send({
+        result: 'success',
+        vehicle: device
+      })
+    }
+  } catch (err) {
+    await fail(err)
+  }
+}
+
+export const submitVehicleEvent = async (req: AgencyApiRequest, res: AgencyApiResponse) => {
+  const { device_id } = req.params
+
+  const { provider_id } = res.locals
+  const name = providerName(provider_id || 'unknown')
+
+  const recorded = now()
+
+  const event: VehicleEvent = {
+    device_id: req.params.device_id,
+    provider_id: res.locals.provider_id,
+    event_type: lower(req.body.event_type) as VEHICLE_EVENT,
+    event_type_reason: lower(req.body.event_type_reason) as VEHICLE_REASON,
+    telemetry: req.body.telemetry ? { ...req.body.telemetry, provider_id: res.locals.provider_id } : null,
+    timestamp: req.body.timestamp,
+    trip_id: req.body.trip_id,
+    recorded,
+    telemetry_timestamp: undefined, // added for diagnostic purposes
+    service_area_id: null // added for diagnostic purposes
+  }
+
+  if (event.telemetry) {
+    event.telemetry_timestamp = event.telemetry.timestamp
+  }
+
+  async function success() {
+    function fin() {
+      res.status(201).send({
+        result: 'success',
+        recorded,
+        device_id,
+        status: EVENT_STATUS_MAP[event.event_type]
+      })
+    }
+    const delta = now() - recorded
+
+    if (delta > 100) {
+      await log.info(name, 'post event took', delta, 'ms')
+      fin()
+    } else {
+      fin()
+    }
+  }
+
+  /* istanbul ignore next */
+  async function fail(err: Error | Partial<{ message: string }>): Promise<void> {
+    const message = err.message || String(err)
+    if (message.includes('duplicate')) {
+      await log.info(name, 'duplicate event', event.event_type)
+      res.status(409).send({
+        error: 'duplicate_event',
+        error_description: 'an event with this device_id and timestamp has already been received'
+      })
+    } else if (message.includes('not found') || message.includes('unregistered')) {
+      await log.info(name, 'event for unregistered', event.device_id, event.event_type)
+      res.status(400).send({
+        error: 'unregistered',
+        error_description: 'the specified device_id has not been registered'
+      })
+    } else {
+      await log.error('post event fail:', event, message)
+      res.status(500).send(new ServerError())
+    }
+  }
+
+  async function finish() {
+    if (event.telemetry) {
+      event.telemetry.recorded = recorded
+      await writeTelemetry(event.telemetry)
+      await success()
+    } else {
+      await success()
+    }
+  }
+
+  // TODO switch to cache for speed?
+  try {
+    const device = await db.readDevice(event.device_id, provider_id)
+    try {
+      await cache.readDevice(event.device_id)
+    } catch (err) {
+      await Promise.all([cache.writeDevice(device), stream.writeDevice(device)])
+      log.info('Re-adding previously deregistered device to cache', err)
+    }
+    if (event.telemetry) {
+      event.telemetry.device_id = event.device_id
+    }
+    const failure = (await badEvent(event)) || (event.telemetry ? badTelemetry(event.telemetry) : null)
+    // TODO unify with fail() above
+    if (failure) {
+      log.info(name, 'event failure', failure, event)
+      return res.status(400).send(failure)
+    }
+
+    // make a note of the service area
+    event.service_area_id = getServiceArea(event)
+
+    // database write is crucial; failures of cache/stream should be noted and repaired
+    const recorded_event = await db.writeEvent(event)
+    try {
+      await Promise.all([cache.writeEvent(recorded_event), stream.writeEvent(recorded_event)])
+      await finish()
+    } catch (err) {
+      await log.warn('/event exception cache/stream', err)
+      await finish()
+    }
+  } catch (err) {
+    await fail(err)
+  }
+}
+
+export const submitVehicleTelemetry = async (req: AgencyApiRequest, res: AgencyApiResponse) => {
+  const start = Date.now()
+
+  const { data } = req.body
+  const { provider_id } = res.locals
+  if (!provider_id) {
+    res.status(400).send({
+      error: 'bad_param',
+      error_description: 'bad or missing provider_id'
+    })
+    return
+  }
+  const name = providerName(provider_id)
+  const failures: string[] = []
+  const valid: Telemetry[] = []
+
+  const recorded = now()
+  const p: Promise<Device | DeviceID[]> =
+    data.length === 1 && isUUID(data[0].device_id)
+      ? db.readDevice(data[0].device_id, provider_id)
+      : db.readDeviceIds(provider_id)
+  try {
+    const deviceOrDeviceIds = await p
+    const deviceIds = Array.isArray(deviceOrDeviceIds) ? deviceOrDeviceIds : [deviceOrDeviceIds]
+    for (const item of data) {
+      // make sure the device exists
+      const { gps } = item
+      const telemetry: Telemetry = {
+        device_id: item.device_id,
+        provider_id,
+        timestamp: item.timestamp,
+        charge: item.charge,
+        gps: {
+          lat: gps.lat,
+          lng: gps.lng,
+          altitude: gps.altitude,
+          heading: gps.heading,
+          speed: gps.speed,
+          accuracy: gps.hdop,
+          satellites: gps.satellites
+        },
+        recorded
+      }
+
+      const bad_telemetry: ErrorObject | null = badTelemetry(telemetry)
+      if (bad_telemetry) {
+        const msg = `bad telemetry for device_id ${telemetry.device_id}: ${bad_telemetry.error_description}`
+        // append to failure
+        failures.push(msg)
+      } else if (!deviceIds.some(item2 => item2.device_id === telemetry.device_id)) {
+        const msg = `device_id ${telemetry.device_id}: not found`
+        failures.push(msg)
+      } else {
+        valid.push(telemetry)
+      }
+    }
+
+    if (valid.length) {
+      const recorded_telemetry = await writeTelemetry(valid)
+      const delta = Date.now() - start
+      if (delta > 300) {
+        log.info(
+          name,
+          'writeTelemetry',
+          valid.length,
+          `(${recorded_telemetry.length} unique)`,
+          'took',
+          delta,
+          `ms (${Math.round((1000 * valid.length) / delta)}/s)`
+        )
+      }
+      if (recorded_telemetry.length) {
+        res.status(201).send({
+          result: `telemetry success for ${valid.length} of ${data.length}`,
+          recorded: now(),
+          unique: recorded_telemetry.length,
+          failures
+        })
+      } else {
+        await log.info(name, 'no unique telemetry in', data.length, 'items')
+        res.status(400).send({
+          error: 'invalid_data',
+          error_description: 'none of the provided data was unique',
+          result: 'no new valid telemetry submitted',
+          unique: 0
+        })
+      }
+    } else {
+      const body = `${JSON.stringify(req.body).substring(0, 128)} ...`
+      const fails = `${JSON.stringify(failures).substring(0, 128)} ...`
+      log.info(name, 'no valid telemetry in', data.length, 'items:', body, 'failures:', fails)
+      res.status(400).send({
+        error: 'invalid_data',
+        error_description: 'none of the provided data was valid',
+        result: 'no valid telemetry submitted',
+        failures
+      })
+    }
+  } catch (err) {
+    res.status(400).send({
+      error: 'invalid_data',
+      error_description: 'none of the provided data was valid',
+      result: 'no valid telemetry submitted',
+      failures: [`device_id ${data[0].device_id}: not found`]
+    })
   }
 }
